requirements.txt
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
anytree==2.12.1
asyncpg==0.28.0
bcrypt==4.0.1
cffi==1.16.0
click==8.1.7
cryptography==41.0.5
dnspython==2.4.2
email-validator==2.0.0.post2
fastapi==0.104.1
fastapi-users==12.1.2
fastapi-users-db-sqlalchemy==6.0.1
greenlet==3.0.1
h11==0.14.0
idna==3.4
jsonify==0.5
makefun==1.15.1
Mako==1.2.4
MarkupSafe==2.1.3
passlib==1.7.4
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
PyJWT==2.8.0
python-dotenv==1.0.0
python-multipart==0.0.6
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
starlette==0.27.0
typing_extensions==4.8.0
uvicorn==0.23.2
